% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{sum_stats}
\alias{sum_stats}
\title{Summarize the input}
\usage{
sum_stats(x)
}
\arguments{
\item{x}{A vector containing a distribution of real numbers}
}
\value{
A string containing the summary statistics of interest
}
\description{
Summarize the input
}
\examples{
sum_stats(c(1:10))
sum_stats(c(1:100))
}
